rename directory for clarity
