+++ src/common/helpers/requestMetaData/approvedIdentifiersTableParser.ts
